examples/training/mixtral/requirements.txt
transformers==4.36.2
pytorch-lightning==2.1.0
regex
tensorboard
datasets
sentencepiece
